+++ docs/Systems/Tailwind/ClassnameReferences/BorderWidth.stories.tsx
@@ -2,11 +2,12 @@ import React from "react"
 import { Story } from "@storybook/react"
 import { Divider } from "@kaizen/draft-divider"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
-import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
+import { StoryWrapper } from "../../../../storybook/components/StoryWrapper"
+import { CATEGORIES } from "../../../../storybook/constants"
 import styles from "./styles.module.scss"
 
 export default {
-  title: "Tailwind/Borders/Border Width",
+  title: `${CATEGORIES.tailwind}/Classname References/Borders/Border Width`,
   component: <div>Hello</div>,
   parameters: {
     docs: {