+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.10...@kaizen/draft-guidance-block@6.1.11) (2023-02-09)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.9...@kaizen/draft-guidance-block@6.1.10) (2023-02-01)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.8...@kaizen/draft-guidance-block@6.1.9) (2023-01-31)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.7...@kaizen/draft-guidance-block@6.1.8) (2023-01-30)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [6.1.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.6...@kaizen/draft-guidance-block@6.1.7) (2023-01-24)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block