+++ draft-packages/hero-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.2.14...@kaizen/draft-hero-card@2.2.15) (2023-01-30)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
 ## [2.2.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.2.13...@kaizen/draft-hero-card@2.2.14) (2023-01-18)
 
 **Note:** Version bump only for package @kaizen/draft-hero-card