+++ draft-packages/menu/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.28](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.8.27...@kaizen/draft-menu@4.8.28) (2023-02-09)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+## [4.8.27](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.8.26...@kaizen/draft-menu@4.8.27) (2023-02-01)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+## [4.8.26](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.8.25...@kaizen/draft-menu@4.8.26) (2023-01-30)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
 ## [4.8.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.8.24...@kaizen/draft-menu@4.8.25) (2023-01-23)
 
 **Note:** Version bump only for package @kaizen/draft-menu