+++ draft-packages/popover/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@5.2.23...@kaizen/draft-popover@5.2.24) (2023-02-09)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
+## [5.2.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@5.2.22...@kaizen/draft-popover@5.2.23) (2023-02-01)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
+## [5.2.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@5.2.21...@kaizen/draft-popover@5.2.22) (2023-01-30)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
 ## [5.2.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@5.2.20...@kaizen/draft-popover@5.2.21) (2023-01-23)
 
 **Note:** Version bump only for package @kaizen/draft-popover