+++ draft-packages/table/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@5.6.8...@kaizen/draft-table@5.6.9) (2023-02-09)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [5.6.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@5.6.7...@kaizen/draft-table@5.6.8) (2023-02-06)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [5.6.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@5.6.6...@kaizen/draft-table@5.6.7) (2023-02-01)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [5.6.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@5.6.5...@kaizen/draft-table@5.6.6) (2023-01-30)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
 ## [5.6.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@5.6.4...@kaizen/draft-table@5.6.5) (2023-01-23)
 
 **Note:** Version bump only for package @kaizen/draft-table