+++ packages/responsive/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/responsive@1.8.4...@kaizen/responsive@1.8.5) (2023-01-30)
+
+**Note:** Version bump only for package @kaizen/responsive
+
+
+
+
+
 ## [1.8.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/responsive@1.8.3...@kaizen/responsive@1.8.4) (2023-01-18)
 
 **Note:** Version bump only for package @kaizen/responsive