+++ packages/tailwind/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.3.1...@kaizen/tailwind@0.3.2) (2023-01-31)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
+## [0.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.3.0...@kaizen/tailwind@0.3.1) (2023-01-31)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
+# [0.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.2.7...@kaizen/tailwind@0.3.0) (2023-01-30)
+
+
+### Features
+
+* (tailwind) Add prettier package ([#3175](https://github.com/cultureamp/kaizen-design-system/issues/3175)) ([077b1b3](https://github.com/cultureamp/kaizen-design-system/commit/077b1b3ed69f7490859aa54597bcafe9ac8440c5))
+
+
+
+
+
 ## [0.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.2.6...@kaizen/tailwind@0.2.7) (2023-01-18)
 
 **Note:** Version bump only for package @kaizen/tailwind